Upgrade UI theme, add dark mode & integrate new dependencies
This commit integrates the NextUI library and adds "clsx" and "prettier" to the devDependencies for enhanced coding standards and efficient concatenation of classnames. Changes are made to the Tailwind.config.ts to include nextui styles in dark mode. Adjustments in app/layout.tsx and app/providers.tsx enable the use of these UI and dark-mode changes across the app. Lastly, a .npmrc file is added to setup the public hoisting of @nextui-org/theme.