Add drag-and-drop functionality with react-dnd library
This commit introduces the 'react-dnd' and 'react-dnd-html5-backend' libraries to the project to implement drag-and-drop functionality. New files 'app/components/cards.tsx' and 'app/constants/index.ts' was added to handle icon cards that will use this functionality. These changes are crucial to enable user interactions with icon cards within our application. This also includes updates to 'pnpm-lock.yaml', which are necessary to maintain consistency in our package dependencies.