Refactor search implementation and update UI
In this commit, the search.tsx file has been moved to the components directory. Inside this file, the structure of the SearchInput component has been modified to encapsulate a SearchIcon and additional text. This provides a more interactive search field, improving the user experience when inputting their queries. Also, the nav component has been commented out in the layout file and replaced with the newly defined Search component in the page file. Exports of some components have been altered as well to comply with the latest changes. Lastly, some of the CSS attributes in the tailwind.config.ts file have been commented out to reflect all the styling updates made.