Update search component with live filtering and svg rendering
Added live search filtering for icons and the react-inlinesvg package, which is used for better rendering of SVG files. Now, the search component can filter the icons in real-time according to the inputted term, improving the component's interactivity. Additionally, SVG files are now rendered using react-inlinesvg which offers superior compatibility with React than dangerouslySetInnerHTML. This change enhances security and performance in rendering SVGs. The package was added to dependencies in package.json and pnpm-lock.yaml.