Improve search functionality and SVG representation
Updated the search component so that searching happens live with every change to the search input, as opposed to on form submission, improving user friendliness. Also, SVG files are now handled more elegantly. They are provided with a rounded background and have their size increased for better visibility. Their labels are presented below each SVG for easier identification. This greatly improves the user's