Add Zustand package for global state management
Zustand package has been added to package.json and integrated into the app in a new file store/index.ts for global state management.
This package was chosen over Redux and Context API for its simplicity and minimalism. In this commit, Zustand is being used to manage a simple counter state. The newly created Zustand store now hosts the state and methods for incrementing and decrementing the counter. We'll be expanding and refactoring this also to other existing global states in our next commits.