Improve icon-card UI and refactor state management
Added padding to the icons display section in search.tsx for better spacing between the edge of the window and the icons. Changed the store currently used for state management to useIconStore to match with the changes made in the state management file. Furthermore, the number of instances for an icon is now displayed on top of the icon, improving UX by providing a clear visual cue of icon selection.

Refactored the state management in the store/index.ts, defining an IconState interface and use it to create the global state useIconStore, making the state management more specific and thus improving readability. Removed unnecessary console logs cleaning up the code further.