Refactor useMemo and Search components within search.tsx
Refactored the useMemo hook out of the Search functional component to a new function for better segregation of concerns and code readability. The new function is named useFilteredMemoIconsList and it takes in a 'term' argument to perform a filter operation on icons based on the term value.

The previous approach of defining useMemo within the Search component made it mentally taxing to understand the operation and it also cluttered up the Search component. The new approach makes both useMemo operation and Search component cleaner and easier to understand.