Refactor DropSection as separate component
Moved the `DropSection` function from the `search.tsx` file to a new `drop-section.tsx` file. This was done to increase modularity and keep each component in its own file. This makes it easier to understand the codebase, isolate component-related issues and re-use the `DropSection` component in other areas of the application if needed.